Rename the variant factory to epicVariant
In preparation for adding a bannerVariant factory.
